+++ packages/zowe-explorer/__tests__/__unit__/extension.unit.test.ts
@@ -243,6 +243,7 @@ async function createGlobalMocks() {
             "zowe.jobs.cancelJob",
             "zowe.jobs.sortBy",
             "zowe.manualPoll",
+            "zowe.editHistory",
             "zowe.updateSecureCredentials",
             "zowe.promptCredentials",
             "zowe.profileManagement",