+++ packages/zowe-explorer/src/job/ZosJobsProvider.ts
@@ -628,6 +628,31 @@ export class ZosJobsProvider extends ZoweTreeProvider implements IZoweTree<IZowe
         return;
     }
 
+    public removeSearchHistory(name: string): void {
+        ZoweLogger.trace("ZosJobsProvider.removeSearchHistory called.");
+        this.mHistory.removeSearchHistory(name);
+    }
+
+    public resetSearchHistory(): void {
+        ZoweLogger.trace("ZosJobsProvider.resetSearchHistory called.");
+        this.mHistory.resetSearchHistory();
+    }
+
+    public getSessions(): string[] {
+        ZoweLogger.trace("DatasetTree.getSessions called.");
+        return this.mHistory.getSessions();
+    }
+
+    public getFileHistory(): string[] {
+        ZoweLogger.trace("DatasetTree.getFileHistory called.");
+        return this.mHistory.getFileHistory();
+    }
+
+    public getFavorites(): string[] {
+        ZoweLogger.trace("DatasetTree.getFavorites called.");
+        return this.mHistory.readFavorites();
+    }
+
     public async getUserJobsMenuChoice(): Promise<FilterItem | undefined> {
         ZoweLogger.trace("ZosJobsProvider.getUserJobsMenuChoice called.");
         const items: FilterItem[] = this.mHistory